Consolidate actions files into single actions.ts file for payment-provisioning component
