+++ packages/webcomponents/src/components/business-forms/payment-provisioning/payment-provisioning.tsx
@@ -3,8 +3,7 @@ import { LoadingSpinner } from '../../form/utils';
 import { BusinessFormClickActions, BusinessFormClickEvent, BusinessFormSubmitEvent } from '../utils/business-form-types';
 import JustifiAnalytics from '../../../api/Analytics';
 import { ComponentError, ComponentErrorCodes, ComponentErrorSeverity } from '../../../api/ComponentError';
-import { makeGetBusiness } from './business-actions';
-import { makePostProvisioning } from './provision-actions';
+import { makeGetBusiness, makePostProvisioning } from './payment-provisioning-actions';
 import { BusinessService } from '../../../api/services/business.service';
 import { ProvisionService } from '../../../api/services/provision.service';
 import { checkProvisioningStatus } from '../utils/helpers';
